+++ external.php
@@ -1,0 +1,4 @@
+// TODO
+
+// I will be added the ability for this log viewer to run outside of the regular wordpress environment.  
+// That would allow for easy debugging when a fatal error occures and wordpress won't load.